Move them to different package and rename
