--- lib/src/main/java/com/otaliastudios/transcoder/internal/opengl/OutputSurface.java
@@ -1,163 +1,0 @@
-/*
- * Copyright (C) 2013 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-// from: https://android.googlesource.com/platform/cts/+/lollipop-release/tests/tests/media/src/android/media/cts/OutputSurface.java
-// blob: fc8ad9cd390c5c311f015d3b7c1359e4d295bc52
-// modified: change TIMEOUT_MS from 500 to 10000
-package com.otaliastudios.transcoder.internal.opengl;
-import android.graphics.SurfaceTexture;
-import android.opengl.EGL14;
-import android.opengl.EGLConfig;
-import android.opengl.EGLContext;
-import android.opengl.EGLDisplay;
-import android.opengl.EGLSurface;
-import android.view.Surface;
-
-import com.otaliastudios.transcoder.internal.Logger;
-
-/**
- * Holds state associated with a Surface used for MediaCodec decoder output.
- * <p>
- * The (width,height) constructor for this class will prepare GL, create a SurfaceTexture,
- * and then create a Surface for that SurfaceTexture.  The Surface can be passed to
- * MediaCodec.configure() to receive decoder output.  When a frame arrives, we latch the
- * texture with updateTexImage, then render the texture with GL to a pbuffer.
- * REMOVED THIS
- * <p>
- * The no-arg constructor skips the GL preparation step and doesn't allocate a pbuffer.
- * Instead, it just creates the Surface and SurfaceTexture, and when a frame arrives
- * we just draw it on whatever surface is current.
- * <p>
- * By default, the Surface will be using a BufferQueue in asynchronous mode, so we
- * can potentially drop frames.
- */
-public class OutputSurface implements SurfaceTexture.OnFrameAvailableListener {
-    private static final String TAG = "OutputSurface";
-    private static final Logger LOG = new Logger(TAG);
-    private EGLDisplay mEGLDisplay = EGL14.EGL_NO_DISPLAY;
-    private EGLContext mEGLContext = EGL14.EGL_NO_CONTEXT;
-    private EGLSurface mEGLSurface = EGL14.EGL_NO_SURFACE;
-    private SurfaceTexture mSurfaceTexture;
-    private Surface mSurface;
-    private final Object mFrameSyncObject = new Object();     // guards mFrameAvailable
-    private boolean mFrameAvailable;
-    private TextureRender mTextureRender;
-
-    /**
-     * Creates an OutputSurface using the current EGL context (rather than establishing a
-     * new one).  Creates a Surface that can be passed to MediaCodec.configure().
-     */
-    public OutputSurface() {
-        mTextureRender = new TextureRender();
-        mTextureRender.surfaceCreated();
-        // Even if we don't access the SurfaceTexture after the constructor returns, we
-        // still need to keep a reference to it.  The Surface doesn't retain a reference
-        // at the Java level, so if we don't either then the object can get GCed, which
-        // causes the native finalizer to run.
-        LOG.v("textureID=" + mTextureRender.getTextureId());
-        mSurfaceTexture = new SurfaceTexture(mTextureRender.getTextureId());
-        // This doesn't work if OutputSurface is created on the thread that CTS started for
-        // these test cases.
-        //
-        // The CTS-created thread has a Looper, and the SurfaceTexture constructor will
-        // create a Handler that uses it.  The "frame available" message is delivered
-        // there, but since we're not a Looper-based thread we'll never see it.  For
-        // this to do anything useful, OutputSurface must be created on a thread without
-        // a Looper, so that SurfaceTexture uses the main application Looper instead.
-        //
-        // Java language note: passing "this" out of a constructor is generally unwise,
-        // but we should be able to get away with it here.
-        mSurfaceTexture.setOnFrameAvailableListener(this);
-        mSurface = new Surface(mSurfaceTexture);
-    }
-
-    /**
-     * Discard all resources held by this class, notably the EGL context.
-     */
-    public void release() {
-        if (mEGLDisplay != EGL14.EGL_NO_DISPLAY) {
-            EGL14.eglDestroySurface(mEGLDisplay, mEGLSurface);
-            EGL14.eglDestroyContext(mEGLDisplay, mEGLContext);
-            EGL14.eglReleaseThread();
-            EGL14.eglTerminate(mEGLDisplay);
-        }
-        mSurface.release();
-        // this causes a bunch of warnings that appear harmless but might confuse someone:
-        //  W BufferQueue: [unnamed-3997-2] cancelBuffer: BufferQueue has been abandoned!
-        //mSurfaceTexture.release();
-        mEGLDisplay = EGL14.EGL_NO_DISPLAY;
-        mEGLContext = EGL14.EGL_NO_CONTEXT;
-        mEGLSurface = EGL14.EGL_NO_SURFACE;
-        mTextureRender = null;
-        mSurface = null;
-        mSurfaceTexture = null;
-    }
-
-    /**
-     * Returns the Surface that we draw onto.
-     *
-     * @return the output surface
-     */
-    public Surface getSurface() {
-        return mSurface;
-    }
-
-    /**
-     * Latches the next buffer into the texture.  Must be called from the thread that created
-     * the OutputSurface object, after the onFrameAvailable callback has signaled that new
-     * data is available.
-     */
-    public void awaitNewImage() {
-        final int TIMEOUT_MS = 10000;
-        synchronized (mFrameSyncObject) {
-            while (!mFrameAvailable) {
-                try {
-                    // Wait for onFrameAvailable() to signal us.  Use a timeout to avoid
-                    // stalling the test if it doesn't arrive.
-                    mFrameSyncObject.wait(TIMEOUT_MS);
-                    if (!mFrameAvailable) {
-                        // TODO: if "spurious wakeup", continue while loop
-                        throw new RuntimeException("Surface frame wait timed out");
-                    }
-                } catch (InterruptedException ie) {
-                    throw new RuntimeException(ie);
-                }
-            }
-            mFrameAvailable = false;
-        }
-        // Latch the data.
-        mTextureRender.checkGlError("before updateTexImage");
-        mSurfaceTexture.updateTexImage();
-    }
-
-    /**
-     * Draws the data from SurfaceTexture onto the current EGL surface.
-     */
-    public void drawImage() {
-        mTextureRender.drawFrame(mSurfaceTexture);
-    }
-
-    @Override
-    public void onFrameAvailable(SurfaceTexture st) {
-        LOG.v("New frame available");
-        synchronized (mFrameSyncObject) {
-            if (mFrameAvailable) {
-                throw new RuntimeException("mFrameAvailable already set, frame could be dropped");
-            }
-            mFrameAvailable = true;
-            mFrameSyncObject.notifyAll();
-        }
-    }
-}
